# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JobConnect AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List / search jobs
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: experienceLevel
          in: query
          schema:
            type: string
        - name: salaryMin
          in: query
          schema:
            type: integer
        - name: salaryMax
          in: query
          schema:
            type: integer
        - name: skills
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
        - name: isRemote
          in: query
          schema:
            type: boolean
        - name: isInternship
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job (recruiter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"

    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update job (recruiter)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete job (recruiter)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /jobs/{id}/save:
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save / unsave a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveToggleResponse"

  # APPLICATIONS
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user resumes
      responses:
        "200":
          description: Resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"

    post:
      operationId: createResume
      tags: [resumes]
      summary: Create/save a resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

    put:
      operationId: updateResume
      tags: [resumes]
      summary: Update resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    SaveToggleResponse:
      type: object
      required: [saved]
      properties:
        saved:
          type: boolean
...
    Job:
      type: object
      required: [id, title, companyId, companyName, location, type, postedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        companyLogoUrl:
          type: ["string", "null"]
        location:
          type: string
        type:
          type: string
          enum: [full-time, part-time, contract, internship, remote]
        experienceLevel:
          type: string
          enum: [entry, mid, senior, lead, executive]
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        skills:
          type: array
          items:
            type: string
        isRemote:
          type: boolean
        isActive:
          type: boolean
        applicantCount:
          type: integer
        savedCount:
          type: integer
        postedAt:
          type: string
        isSaved:
          type: boolean
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            description:
              type: string
            requirements:
              type: array
              items:
                type: string
            responsibilities:
              type: array
              items:
                type: string
            benefits:
              type: array
              items:
                type: string
            company:
              $ref: "#/components/schemas/Company"
    JobInput:
      type: object
      required: [title, companyId, location, type, description]
      properties:
        title:
          type: string
        companyId:
          type: integer
        location:
          type: string
        type:
          type: string
        experienceLevel:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        skills:
          type: array
          items:
            type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        responsibilities:
          type: array
          items:
            type: string
        benefits:
          type: array
          items:
            type: string
        isRemote:
          type: boolean
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Company:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        openPositions:
          type: integer
        isVerified:
          type: boolean
...
    Resume:
      type: object
      required: [id, userId, title, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        templateId:
          type: ["string", "null"]
        content:
          type: object
          additionalProperties: true
        atsScore:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ResumeInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        templateId:
          type: string
        content:
          type: object
          additionalProperties: true

    # AI
...
    Notification:
      type: object
      required: [id, userId, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [application_update, interview_invite, job_alert, system, message]
        message:
          type: string
        isRead:
          type: boolean
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string

    # DASHBOARDS
...