# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * JobConnect AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';
import type { Job } from './job';

export type JobDetail = Job & {
  description?: string;
  requirements?: string[];
  responsibilities?: string[];
  benefits?: string[];
  company?: Company;
};